+++ rental/src/main/java/pl/edu/agh/rental/rental/dto/RentalData.java
@@ -16,5 +16,5 @@ public record RentalData(
         Long rentalCurrentTime,
 
         @NotNull
-        Double rentalCost
+        Long rentalCost
 ) { }
